joint-segmentation-depth/requirements.txt
# PyTorch and related
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Segmentation models
segmentation-models-pytorch>=0.3.3
transformers>=4.30.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
lxml>=4.9.0

# Experiment tracking
wandb>=0.15.0

# Development tools
ipython>=8.0.0
jupyter>=1.0.0
pytest>=7.0.0

# Optional: For faster data loading
# opencv-python-headless>=4.8.0  # Use instead of opencv-python for server environments